Drivers: CholQRCP and QB-backed SVD (#15)
- Addition of driver routines -
[RSVD](https://github.com/BallisticLA/RandLAPACK/blob/max-benchmark/src/drivers/rsvd.cc)
and
[CholQRCP](https://github.com/BallisticLA/RandLAPACK/blob/max-benchmark/src/drivers/cholqrcp.cc)
(with related tests).
- Reworked benchmarking infrastructure - benchmarks now need to be
compiled separately from RandLAPACK itself.
- Addition of various benchmarks for CholQRCP.
- Expansion of the set of utility functions (primarily added things
required by CholQRCP).

**TODOs, to handle after this PR**
- Need javadoc style comments for Doxygen
- At the moment, benchmarking code saves results into specified
locations. If the required folder structure is absent, files will
not be saved. Need to rework that.
- Older benchmarks
([Orth_speed](https://github.com/BallisticLA/RandLAPACK/blob/max-benchmark/benchmark/Orth_speed.cc)
and
[QB_cond_nums](https://github.com/BallisticLA/RandLAPACK/blob/max-benchmark/benchmark/QB_cond_nums.cc))
may need to be revised.
- Utility functions for extracting lower- and upper-triangular portions
of the matrix should be written in two formats: when the initial data is
modified and when entries are copied into a new buffer.
- Some benchmarks use tuples instead of vectors - needs fixing. 
- In CholQRCP, a runtime logging structure was introduced. We may want
to apply similar logic to other driver and some computational routines.
- Need to write tests for utility routines.
- I am not certain if the
[col_swap](https://github.com/BallisticLA/RandLAPACK/blob/max-benchmark/src/comps/util.cc#L130)
routine is written in the best way possible.
- Need to rework algorithm classes to have one algorithm per class.

Co-authored-by: Melnichenko Max <thrcn@TeachRaccoon>