RBKI (#61)
This PR adds a new driver algorithm - RBKI (Randomized Block Krylov
Iteration) (formal name subject to change).
RBKI algorithm is a method for finding truncated SVD based on block
Krylov iterations.
RBKI is applicable for matrices with a slowly-decaying spectrum; such
cases are canonically considered ''hard`` for a standard randomized SVD.

This algorithm is a version of Algorithm A.1 from
https://arxiv.org/pdf/2306.12418.pdf
The main difference is in the fact that an economy SVD is performed only
once at the very end
of the algorithm run and that the termination criteria are not based on
singular vector residual evaluation.
Instead, the scheme terminates if:
1. ||R||_F > sqrt(1 - eps^2) ||A||_F, which ensures that we've exhausted
all vectors and doing more
iterations would bring no benefit or that ||A - hat(A)||_F < eps *
||A||_F.
2. Stop if the bottom right entry of R or S is numerically close to zero
(up to the square root of machine eps).
        
The main cost of this algorithm comes from large GEMMs with the input
matrix A.
The algorithm optionally times all its subcomponents through a
user-defined 'timing' parameter.
Some things about the algorithm to keep in mind:
1. The need to iteratively allocate space for Krylov subspace buffers
(Y_od, X_ev), as well as for the matrices that are to
be decomposed via SVD at the end of the alg (R, S). Additional
allocation is done via `realloc()`, followed up by `memset()`
call to ensure that the newly allocated space has been zeroed out.
2. Matrix R needs to be kept in a transposed format at all times, as the
final decomposition via SVD is to be done on R'.
3. Reorthogonalization steps are a must. These require additional (n +
b_sz) by b_sz and n by b_sz computational buffers.
4. 'Q-factors' from QR factorization at every iteration form the Krylov
subspace -> we need explicit versions of those. Hence,
         using `ungqr()` is required.
5. Even though the algorithm accepts a "tolerance" parameter, we can
control its termination using the "max_iters"
          parameter.
 

In addition to the RBKI, the following is also introduced in this PR:
1. Switch from using `std::vector` to pointers in the `/misc/rl_gen.hh`
and `/misc/rl_util.hh`
2. Addition of `process_input_mat()` functionality in `/misc/rl_gen.hh`
- allows to read a given input matrix into RandLAPACK (without knowing
its size in advance) from a file.
3. Gemm vs ormqr speed benchmark. 
4. RBKI runtime breakdown benchmark - assesses the time taken by each
subcomponent of RBKI.
5. RBKI speed comparison benchmark - technically only runs RBKI, but has
an option to run SVD (`gesdd()`) to be compared against RBKI (direct SVD
is WAY slower than RBKI). The user is required to provide a matrix file
to be read, set min and max numbers of large gemms (Krylov iterations)
that the algorithm is allowed to perform min and max block sizes that
RBKI is to use; furthermore, the user is to provide a 'custom rank'
parameter (number of singular vectors to approximate by RBKI). The
benchmark outputs the basic data of a given run, as well as the RBKI
runtime and singular vector residual error, which is computed as
"sqrt(||AV - SU||^2_F + ||A'U - VS||^2_F / sqrt(custom_rank)" (for
"custom rank" singular vectors and values).